Fixed test failure
